requirements.txt
fastapi
python-dotenv
pyodbc
uvicorn
gunicorn
